v0.0.7(UPDATE: INVENTORY)

+++ mini-game/source/modules/products/products.h
@@ -1,20 +1,16 @@
-//#ifndef PRODUCTS_H
-//#define PRODUCTS_H
-
 #include <iostream>
 #include <string>
 #include <vector>
 using namespace std;
 
+// Характеристики товаров
 class Product
 {
 public:
     int id;
     string name;
     double price;
-    int count;
+    double count;
 };
 
 extern vector <Product> products;
-
-//#endif
